fix(coverage): address PR #19 review comments
- Thread safety: @Synchronized on markCovered() and reset() to guard
  the compound read-check-write on coveredCells and coveredCount
- Hilt binding: add CoverageModule providing CoverageTracker as @Singleton
- DevicePose: change Float fields to Double and add KDoc with units/ranges
- isBelowWarningThreshold: accept threshold param (default in CoverageTracker
  companion) so callers can override; removes UX policy from domain interface
- SphereRegion: add TODO(#22) directing migration to DevicePose
- Tests: update DevicePose literals to Double, replace SphereCoverageTracker
  .WARNING_THRESHOLD with CoverageTracker.DEFAULT_WARNING_THRESHOLD, add
  grid cell position assertion, add extreme/out-of-range pitch boundary tests,
  add custom-threshold test

+++ app/src/main/java/com/volta/app/domain/model/DevicePose.kt
@@ -1,7 +1,13 @@
 package com.volta.app.domain.model
 
+/**
+ * Device orientation in 3D space, all values in radians.
+ * @param yaw Rotation around the vertical axis. Any value accepted; normalized internally to [0, 2π).
+ * @param pitch Elevation angle in [-π/2, π/2], where -π/2 is straight down and π/2 is straight up.
+ * @param roll Rotation around the forward axis in [-π, π]. Not used for coverage tracking.
+ */
 data class DevicePose(
-    val yaw: Float,
-    val pitch: Float,
-    val roll: Float
+    val yaw: Double,
+    val pitch: Double,
+    val roll: Double
 )
